changed gif data collection, fixed MinRect class issue with VU9P

+++ src/main/AutoPlacement.java
@@ -207,365 +207,6 @@ public class AutoPlacement {
         return replicated;
     }
 
-    /* @deprecated
-       this is a complete toolflow without SLR replication, takes about 1.5 hours - 2 hours
-    */
-    public static void flow_regular() throws IOException {
-        // read config
-        Properties prop = Tool.getProperties();
-        String device = prop.getProperty("device");
-        String part = new Design("name", device).getPartName();
-
-        // Switches
-        final boolean optimization = Boolean.parseBoolean(prop.getProperty("optimization"));
-        final boolean rapidSynth = Boolean.parseBoolean(prop.getProperty("rapidSynth"));
-        final boolean autoPipeline = Boolean.parseBoolean(prop.getProperty("autoPipeline"));
-        final boolean matplotlib_visualize = Boolean.parseBoolean(prop.getProperty("matplotlib_visual"));
-        final boolean vivado_verbose = Boolean.parseBoolean(prop.getProperty("vivado_verbose"));
-        final boolean visualization = Boolean.parseBoolean(prop.getProperty("opt_visual"));
-
-        // experiment config
-        int blocknum = 80; // TODO: automatically determine blocknum
-        String method = prop.getProperty("method");
-
-        // optimization parameters
-        int population = 5;
-        int parents = 20;
-        int children = 50;
-        double crossoverR = 0.98;
-        int x_min = 0;
-        int x_max = 6000; // all columns
-        int y_min = 0;
-        int y_max = 240; // TODO: automatically determine min replicating rectangle
-
-        // set up paths
-        String root = System.getProperty("RAPIDWRIGHT_PATH") + "/";
-        String checkpoint = root + "checkpoint/";
-        String results = root + "result/";
-
-
-        /*  --- find placement solution --- */
-        Map<Integer, List<Site[]>> result = new HashMap<>();
-        String xdc_result = results + "blockNum=" + blocknum + ".xdc";
-
-        if (optimization)
-        {
-            FileWriter fw = new FileWriter(xdc_result);
-            PrintWriter pw = new PrintWriter(fw, true);
-            result = find_solution(
-                    method, blocknum, visualization, device,
-                    population, parents, children, crossoverR,
-                    x_min, x_max, y_min, y_max);
-
-            Tool.write_XDC(result, pw);
-            pw.close();
-        }
-        else // use previous results
-            result = Tool.getMapFromXDC(xdc_result, device);
-        System.out.println("Found Placement Strategy for " + result.size() + " blocks of convolution units");
-
-
-        /* generate full-chip placement */
-        result = populate(result, device);
-        blocknum = 480;
-
-
-        /* placement result visualization */
-        if (matplotlib_visualize) {
-            Tool.matplot_visualize(xdc_result);
-        }
-
-        /* Synthesis */
-        Design d;
-        System.out.println("start synthesis...");
-        if (rapidSynth)
-            d = Vivado.synthesize_with_seed(blocknum, device, 0, part, vivado_verbose);
-        else
-            d = Vivado.synthesize_vivado(blocknum, part, 0, vivado_verbose);
-        d.setAutoIOBuffers(false); // out of context mode
-        System.out.println("synthesis finished");
-
-
-        /* Placement */
-        long start_time = System.nanoTime();
-        System.out.println("Placement Start...");
-        for (Integer index : result.keySet()) {
-            List<Site[]> blockConfig = result.get(index);
-            place_block(d, index, blockConfig);
-        }
-        System.out.println("Site-Routing ...");
-        d.routeSites();
-        long end_time = System.nanoTime();
-
-        System.out.println(">>>-----------------------------------------------");
-        final String s = "RapidWright Hard Block Placement time = " + (end_time - start_time) / 1e9
-                + " s, which is " + (end_time - start_time) / 1e9 / 60 + " min";
-        System.out.println(s);
-        System.out.println(">>>-----------------------------------------------");
-
-        /* Auto Pipeline */
-        if (autoPipeline) {
-            System.out.println(">>>------------ Auto-pipeline ----------<<<");
-            AutoPipeline.rapidsynth_autopipeline(d, result);
-        }
-
-        // output placed DCP file
-        String placedDCPPath = checkpoint + "blockNum=" + blocknum + "_placed.dcp";
-        File file = new File(placedDCPPath);
-        if (file.exists())
-            file.delete();
-        d.writeCheckpoint(placedDCPPath);
-
-        // finish placement and routing with Vivado
-        double freq = Vivado.finishPlacementNRoute(placedDCPPath, blocknum, vivado_verbose);
-        System.out.println("$$$ TIMING RESULT : block num  = " + blocknum + "\t frequency = " + freq/1e6 + "MHz");
-
-    }
-
-    /* @deprecated
-        this is a complete toolflow with SLR replication, takes about 1 hours
-    */
-    public static void flow_SLR() throws IOException {
-        // read config
-        Properties prop = Tool.getProperties();
-        String device = prop.getProperty("device");
-        String part = new Design("name", device).getPartName();
-
-        // Switches
-        final boolean optimization = Boolean.parseBoolean(prop.getProperty("optimization"));
-        final boolean rapidSynth = Boolean.parseBoolean(prop.getProperty("rapidSynth"));
-        final boolean autoPipeline = Boolean.parseBoolean(prop.getProperty("autoPipeline"));
-        final boolean matplotlib_visualize = Boolean.parseBoolean(prop.getProperty("matplotlib_visual"));
-        final boolean vivado_verbose = Boolean.parseBoolean(prop.getProperty("vivado_verbose"));
-        final boolean visualization = Boolean.parseBoolean(prop.getProperty("opt_visual"));
-        final int depth = Integer.parseInt(prop.getProperty("pipelineDepth"));
-
-        MinRect mr = new MinRect(device, 18, 8, 2);
-        int blocknum = mr.getBlocknum();
-        int replication = mr.getReplication();
-        String method = prop.getProperty("method");
-
-        // optimization parameters
-        int population = 5;
-        int parents = 20;
-        int children = 50;
-        double crossoverR = 0.98;
-        int x_min = 0;
-        int x_max = 6000; // use all columns
-        int y_min = 0;
-        int y_max = mr.getYmax();
-
-        // set up paths
-        String root = System.getProperty("RAPIDWRIGHT_PATH") + "/";
-        String checkpoint = root + "checkpoint/";
-        String results = root + "result/";
-
-
-        /*  --- find placement solution --- */
-        Map<Integer, List<Site[]>> result = new HashMap<>();
-        String xdc_result = results + "blockNum=" + blocknum + ".xdc";
-
-        if (optimization)
-        {
-            FileWriter fw = new FileWriter(xdc_result);
-            PrintWriter pw = new PrintWriter(fw, true);
-            result = find_solution(
-                    method, blocknum, visualization, device,
-                    population, parents, children, crossoverR,
-                    x_min, x_max, y_min, y_max);
-
-            Tool.write_XDC(result, pw);
-            pw.close();
-        }
-        else // use previous results
-            result = Tool.getMapFromXDC(xdc_result, device);
-        System.out.println("Found Placement Strategy for " + result.size() + " blocks of convolution units");
-
-        if (matplotlib_visualize)
-            Tool.matplot_visualize(xdc_result);
-
-        /* replicate placement to one SLR */
-        result = AutoPlacement.populateFixed(result, device, replication);
-        blocknum *= mr.getReplication();
-        xdc_result = results + "blockNum=" + blocknum + ".xdc";
-        PrintWriter pr = new PrintWriter(new FileWriter(xdc_result), true);
-        Tool.write_XDC(result, pr);
-        pr.close();
-
-        /* synthesize one SLR */
-        Design d = null;
-        if (rapidSynth)
-            d = Vivado.synthesize_with_seed(
-                    blocknum, device, 0,  part, vivado_verbose);
-        else
-            d = Vivado.synthesize_vivado(blocknum, part, 0, vivado_verbose);
-        System.out.println("One SLR synthesis finished.");
-
-        /* placement and site-routing */
-        long start_time = System.nanoTime();
-        System.out.println("Placement Start...");
-        for (Integer index : result.keySet()) {
-            List<Site[]> blockConfig = result.get(index);
-            AutoPlacement.place_block(d, index, blockConfig);
-        }
-        System.out.println("Site-Routing ...");
-        d.routeSites();
-        long end_time = System.nanoTime();
-        System.out.println(">>>-----------------------------------------------");
-        String s = "RapidWright Hard Block Placement time = " + (end_time - start_time) / 1e9
-                + " s, which is " + (end_time - start_time) / 1e9 / 60 + " min";
-        System.out.println(s);
-        System.out.println(">>>-----------------------------------------------");
-
-        //d = Vivado.legalize_process(d);
-
-        /* pipelining */
-        AutoPipeline.fixed_pipeline(d, depth, blocknum);
-
-        /* write out checkpoint, finish routing with Vivado */
-        String placedDCPPath = checkpoint + "blockNum=" + blocknum + "_placed.dcp";
-        File file = new File(placedDCPPath);
-        if (file.exists())
-            file.delete();
-        d.writeCheckpoint(placedDCPPath);
-        double freq = Vivado.finishPlacementNRoutePBlock(placedDCPPath, blocknum, result, device, vivado_verbose);
-
-        /* read in routed SLR and replicate */
-        String routedSLR = checkpoint + "blockNum=" + blocknum + "_routed.dcp";
-        start_time = System.nanoTime();
-        Design full_chip_routed = Tool.replicateSLR(routedSLR);
-        end_time = System.nanoTime();
-        System.out.println(">>>-----------------------------------------------");
-        s = "SLR Replication time = " + (end_time - start_time) / 1e9
-                + " s, which is " + (end_time - start_time) / 1e9 / 60 + " min";
-        System.out.println(s);
-        System.out.println(">>>-----------------------------------------------");
-        //full_chip_routed.flattenDesign();
-        full_chip_routed.writeCheckpoint(checkpoint + "full-chip_" + device + ".dcp");
-
-        /* report clock frequency */
-        System.out.println("$$$$ frequency =  " + freq/1e6 + " MHz");
-    }
-
-    /* @deprecated
-        this is the debug function for SLR toolflow
-    */
-    public static void flow_SLR_debug() throws IOException {
-        int blockn = 1;
-        int depth = 3;
-        String device = "xcvu11p";
-        String part = new Design("name", device).getPartName();
-
-        /* optimization */
-        Map<Integer, List<Site[]>> result = AutoPlacement.find_solution(
-                "CMA", blockn, false, device,
-                5, 10, 20, 0.98,
-                0, 6000, 0, 240);
-
-        //result = AutoPlacement.populateFixed(result, device, 2);
-        //blockn = blockn * 2;
-
-        /* synthesize one SLR */
-        Design d = Vivado.synthesize_vivado(blockn, part, 0, true);
-
-        /* placement */
-        for (Integer index : result.keySet()) {
-            List<Site[]> blockConfig = result.get(index);
-            AutoPlacement.place_block(d, index, blockConfig);
-        }
-        d.routeSites();
-
-        d = Vivado.legalize_process(d);
-
-        /* pipelining */
-        AutoPipeline.fixed_pipeline(d, depth, blockn);
-
-        String pipelined_dcp = System.getenv("RAPIDWRIGHT_PATH") + "/checkpoint/pipeline.dcp";
-        d.writeCheckpoint(pipelined_dcp);
-
-        /* finish routing */
-//        String tclfile = System.getenv("RAPIDWRIGHT_PATH") + "/tcl/experiment_pp";
-//        PrintWriter tcl = new PrintWriter(new FileWriter(tclfile), true);
-//        tcl.println("open_checkpoint " + pipelined_dcp);
-//        // this is the problem
-//        //tcl.println("create_clock -period 1.000 -waveform {0.000 0.500} [get_nets clk];");
-//        tcl.println("startgroup");
-//        tcl.println("create_pblock {pblock_name.dut}");
-//        tcl.println("resize_pblock {pblock_name.dut} -add " + "CLOCKREGION_X0Y0:CLOCKREGION_X1Y0");
-//        tcl.println("add_cells_to_pblock {pblock_name.dut} -top");
-//        tcl.println("endgroup");
-//        tcl.println("set_property CONTAIN_ROUTING true [get_pblocks pblock_name.dut]");
-//        tcl.println("place_design");
-//
-//        tcl.close();
-//
-//        Vivado.vivado_cmd("vivado -mode tcl -source " + tclfile, true);
-        String placedDCPPath = System.getProperty("RAPIDWRIGHT_PATH") + "/checkpoint/pipeline.dcp";
-        Vivado.finishPlacementNRoutePBlock(placedDCPPath, blockn, result, device, true);
-
-
-    }
-
-    /* @deprecated
-        this is the tool flow with manual placement
-    */
-    public static void flow_manual_placement() throws IOException {
-        // read config
-        Properties prop = Tool.getProperties();
-        String device = prop.getProperty("device");
-        String part = new Design("name", device).getPartName();
-
-        // Switches
-        final boolean rapidSynth = Boolean.parseBoolean(prop.getProperty("rapidSynth"));
-        final boolean vivado_verbose = Boolean.parseBoolean(prop.getProperty("vivado_verbose"));
-
-        // experiment config
-        int blocknum = 480;
-        int depth = 4;
-
-        // set up paths
-        String root = System.getProperty("RAPIDWRIGHT_PATH") + "/";
-        String checkpoint = root + "checkpoint/";
-        String results = root + "result/";
-
-
-        /*  --- read manual placement solution --- */
-        Map<Integer, List<Site[]>> result = Tool.getMapFromXDCRobust(root + "src/verilog/dsp_conv_chip.xdc", device, blocknum);
-
-        /* Synthesis */
-        Design d;
-        System.out.println("start synthesis...");
-        if (rapidSynth)
-            d = Vivado.synthesize_with_seed(blocknum, device, 0, part,  vivado_verbose);
-        else
-            d = Vivado.synthesize_vivado(blocknum, part, depth, vivado_verbose);
-        d.setAutoIOBuffers(false); // out of context mode
-        System.out.println("synthesis finished");
-
-        /* Placement */
-        long start_time = System.nanoTime();
-        System.out.println("Placement Start...");
-        for (Integer index : result.keySet()) {
-            List<Site[]> blockConfig = result.get(index);
-            place_block(d, index, blockConfig);
-        }
-        System.out.println("Site-Routing ...");
-        d.routeSites();
-        long end_time = System.nanoTime();
-
-        System.out.println(">>>-----------------------------------------------");
-        final String s = "RapidWright Hard Block Placement time = " + (end_time - start_time) / 1e9
-                + " s, which is " + (end_time - start_time) / 1e9 / 60 + " min";
-        System.out.println(s);
-        System.out.println(">>>-----------------------------------------------");
-
-        String placed_design = checkpoint + "blockNum=" + blocknum + "_placed.dcp";
-
-        d.writeCheckpoint(placed_design);
-
-        Vivado.finishPlacementNRoutePBlock(placed_design, blocknum, result, device, vivado_verbose);
-    }
-
     /*
         Function: toolflow
         This is the entire toolflow as a demo. The configurations are loaded
@@ -573,14 +214,13 @@ public class AutoPlacement {
     */
     public static void toolflow() throws IOException {
 
-        // TODO: add transfer learning option
         // TODO: add GIF option
-        // TODO: fix autopipeline
 
         // read in configuration
         Properties prop = Tool.getProperties();
         String device = prop.getProperty("device");
         String part = new Design("name", device).getPartName();
+        String placement = prop.getProperty("placement");
 
         // set up flags
         final boolean optimization = Boolean.parseBoolean(prop.getProperty("optimization"));
@@ -588,7 +228,9 @@ public class AutoPlacement {
         final boolean autoPipeline = Boolean.parseBoolean(prop.getProperty("autoPipeline"));
         final boolean vivado_verbose = Boolean.parseBoolean(prop.getProperty("vivado_verbose"));
         final boolean visualization = Boolean.parseBoolean(prop.getProperty("opt_visual"));
+        final boolean generate_gif = Boolean.parseBoolean(prop.getProperty("generate_gif"));
         final int depth = Integer.parseInt(prop.getProperty("pipelineDepth"));
+        boolean placement_exists = new File(placement).exists();
 
         // automatically calc block number, replication times, minimum replicate rectangle
         MinRect mr = new MinRect(device, 18, 8, 2); // hard block number in 1 computation unit
@@ -611,7 +253,7 @@ public class AutoPlacement {
         String xdc_result = results + "blockNum=" + blocknum + ".xdc";
         boolean result_exists = new File(xdc_result).exists();
 
-        if (optimization || !result_exists) {
+        if (optimization || (!result_exists && !placement_exists)) {
             // if we want to rerun optimization or there's no existing xdc files
             FileWriter fw = new FileWriter(xdc_result);
             PrintWriter pw = new PrintWriter(fw, true);
@@ -622,8 +264,10 @@ public class AutoPlacement {
 
             Tool.write_XDC(result, pw);
             pw.close();
-        } else {
+        } else if (placement_exists) {
             // else we use previous results stored in the xdc file
+            result = Tool.getMapFromXDC(placement, device);
+        } else {
             result = Tool.getMapFromXDC(xdc_result, device);
         }
         System.out.println("[RapidLayout] Found Placement Strategy for " + result.size() + " blocks of convolution units");
@@ -661,8 +305,10 @@ public class AutoPlacement {
         System.out.println(">>>-----------------------------------------------<<<");
 
         /* --- pipeline --- */
-        // TODO autopipeline hasn't work yet
-        AutoPipeline.fixed_pipeline(d, depth, blocknum);
+        if (autoPipeline)
+            AutoPipeline.auto_pipeline(d, blocknum, result);
+        else
+            AutoPipeline.fixed_pipeline(d, depth, blocknum);
 
         System.out.println("[RapidLayout] Pipelining finished.");
